requirements.txt
pdf2docx
PyMuPDF
python-docx
